refactoring: moved PickleSerializingMixin to cache and based it on existing dump/load methods
